Finish repath, fix api
