+++ src/module/manager/__init__.py
@@ -1,2 +1,3 @@
 from .eps_complete import FullSeasonGet
 from .renamer import Renamer
+from .repath import set_new_path
