+++ src/module/manager/eps_complete.py
@@ -93,6 +93,10 @@ class FullSeasonGet(DownloadClient):
             self.add_torrent(download)
         logger.info("Completed!")
 
+    def add_subscribe(self, data: BangumiData, link):
+        self.add_rss_feed(link, item_path=data.official_title)
+        self.set_rule(data)
+
 
 if __name__ == '__main__':
     from module.conf import setup_logger
